Create filterIndexersByTimestamp method
